requirements.txt
fastapi==0.115.6
pymongo==4.6.0
python-dotenv==1.0.1
pytz==2024.1
tzlocal==5.2
